retention/requirements.txt
mysqlclient==2.0.3
psycopg2-binary==2.7.6.1
